Modify filter() to just get TC programs
The filter() function used when loading and unloading multi attach
programs was getting both TC and TCX programs because they both have the
same ProgramType's.  I modified it to just get TC programs and renamed
it to get_multi_attach_programs()

I also removed a few more unwrap()s and handled the errors in the process.

Signed-off-by: Andre Fredette <afredette@redhat.com>