bpfman: every load and unload request is logged
Make sure every `load` and `unload` request is logged with the function
name when bpfman is first called and ensure a success or failure is
reported.

Moved most of the contents of `add_program()` and `remove_program()` to
a subfunction so that all errors are captured and existing code could
still easily drop out with a `?`.

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>