doc: update example building to use bpfman image build
The documentation on examples has a section on building the examples
which was still using the `docker build --build-arg ...` style. Switch
it to use the `bpfman build image` command.

Added a paragraph to the `bpfman image generate-build-args` portion of
the CLI Guide to show how to still build with the `docker
build --build-arg ...` style in case it's needed.

Also found a few spots in the documents that were still referencing
the old object names, like `bpf_bpfel.o`. With multi-arch, all our
bytecode object files have the arch in the name, like
`bpf_x86_bpfel.o`.

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>