unpin cargo nightly
Cargo has been pinned to nightly-2024-09-24 for a while. It was
originally pinned because of clippy failures (see #1282). But once those
were resolved, it was decided to leave pinned and move the pin roughly
monthly. Because that hasn't been happening, and it is hard to document
this for new comers to the project, move back to nightly. May move back
to pinning if clippy type errors slow the project down.

Resolves: #1282

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>