image pull failure
Image pull was failing with:

```
$ sudo bpfman load image --image-url quay.io/netobserv/ebpf-bytecode:e705638 \
  --name tcx_ingress_flow_parse tcx --iface eno3 --priority 100 --direction ingress

Error: failed to read db: Database entry quay.io_netobserv_ebpf-bytecode_mainc45136d
ba8fb3c73b87df1a0b125b5b354ae026cdd0ce2740dcf48c3ad54da10 does not exist in tree
"__sled__default":
```

After further investigation, it was discovered that if bpfman is uninstalled and
loaded fresh (empty database), the following error occurs:

```
$ sudo bpfman load image --image-url quay.io/netobserv/ebpf-bytecode:e705638 \
  --name tcx_ingress_flow_parse tcx --iface eno3 --priority 100 --direction ingress

Error: expected `,` or `}` at line 5 column 1
```

After adding additional debug logs, the primary issue was that the image was built
with invalid map labels, there was a comma missing part of the way through the list.
Once this occurred, there was an invalid/empty entry for that image. So subsequent
calls to load checked the database to see if an entry already existed, which it did,
so tried to load from there, and it failed because it really wasn't loaded.

This PR moved all the image pull DB writes to the end of the processing, to make sure
pull succeeded properly before writing to the database. This should fix the "failed
to read db" error. Also added a more informative error for the invalid map label:

```
$ sudo bpfman load image --image-url quay.io/netobserv/ebpf-bytecode:e705638 \
  --name tcx_ingress_flow_parse tcx --iface eno3 --priority 100 --direction ingress

Error: error pulling image, invalid image map label: expected `,` or `}` at line 5 column 1
```

There was some flakiness with he integration tests with image pulling,
but the errors returned from bpfman calls were being lost with
`unwrap()` calls. So updated tests in e2e.rs to not just `unwrap()` the
return value, but log the error if one occurred.

In the image manager code, where the DB calls were al moved to the end
of the function, also moved the DB Flush call to the very end.

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>