chore: Unasync all the things
When we moved bpfman to be a oneshot commandline program instead of
being a daemon we noted that we should run sync and not async given
that a lot of the work we do is synchronous.

The blocking factor here was that sigstore and oci_distribution
required us to use tokio.

This was solved by creating a Tokio current thread runtime when
required and running any async tasks using `block_on`.

In effect, bpfman can remain sync, while using dependnecies that
might require an async runtime like tokio.

This should make debugging bpfman a little easier than it was before and
may hopefully help us figure out any data races we may have - regardless
of the backend db that we're using.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>