chore: Unasync all the things
When we moved bpfman to be a oneshot commandline program instead of
being a daemon we noted that we should run sync and not async given
that a lot of the work we do is synchronous.

The blocking factor here was that sigstore and oci_distribution
required us to use tokio.

This solve this, we introduce a new wrapper called BlockingWrapper
that lets us run:

- Async code from a sync context (i.e oci_utils from bpfman)
- Async code from a sync context from an async context
  (i.e bpfman-rpc -> bpfman -> oci_utils)

In effect, bpfman can remain sync, while using dependnecies that
might require an async runtime like tokio.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>