Fix errors deleting tc and xdp programs in network namespaces
For TC Programs:
When a TC program is attached within a network namespace, switch to
that namespace to detach the TC dispatcher after removing the last
program.

For TC and XDP Programs:
If TC or XDP programs are attached within a network namespace, avoid
attempting to rebuild and reattach the dispatcher if the namespace
has already been deleted when the programs are removed.

Fixes: #1402

Signed-off-by: Andre Fredette <afredette@redhat.com>