chore: Move completion/manpages to hidden cmds
Previously we had decided that manpage and completions generation should
happen in xtask. This required that bpfman/cli/args.rs be completely
standalone such that it can be imported cleanly in xtask.

This hasn't been the case for a while - which means that xtask must
depend on bpfman... Which causes xtask to break if bpfman fails to
compile (i.e you're working on it) which is not ideal.

Moving this to (hidden) subcommands solves the problem, the only
downside being another couple of dependencies in bpfman. This seems like
a reasonable trade-off.

Fixes: #1384

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>