fix(bpfman-rpc): bpfman-rpc performance issues
@Billy99 reported that bpfman-rpc was performing poorly since the
changes to use only sync code inside bpfman.

This was due to the way we were solving for doing async->sync->async
code bridging (yuck!). The long term fix to to add sync client to
rust-oci-client and sigstore libraries. However in the interim the
following approach appears to work.

1. Add a `rt()` function to create a LazyLock tokio multithread runtime
2. Everywhere we need to jump into async code we call `rt()?.block_on()`
3. From the async code, all bpfman functions have to be called from
   `spawn_blocking`.
4. To ensure that is done consistently from the CSI/RPC code we wrap the
   bpfman functions in an AsyncBpfman struct which is also locked with a
   Mutex

The addition of a mutex also protects against any races on acquiring a
database lock.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>