bpfman load should return a summary of IDs
Now that we are loading multiple programs, after a load command it would
be better to just see a summary list of the id's of the programs.
Updated the code such that if one program is being attached, acts as it
did before and the full program is displayed (like a `bpfman get`). If
multiple programs are loaded, they are printed as a list (like a `bpfman
list`).

Resolves: #1426

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>

Stash

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>