support SecurityProfilesOperator v0.9.1
The SecurityProfileOperator (SPO) created a new release, v0.9.0. This
release was pushed to OperatorHub (2/25/25) but has failed to merge.
https://github.com/k8s-operatorhub/community-operators/pull/5799
As part of the release, the SeLinux Profiles CRDs moved from namespace
scoped to cluster scoped. This effects bpfman two ways:
* The SeLinux Profiles bpfman creates in the examples need to remove the
  namespace field.
* In the examples DaemonSet, when deploying with SeLinux, in the
  container Security Context, the yaml points to the SeLinux Profile
  using <SeLinuxProfileName>_<Namespace>.process. With the change to
  namespace scoped CRD, this becomes <SeLinuxProfileName>.process.

This commit adjusts the SeLinux deployment Kustomize files to account
for these changes. Because the SeLinux Profile was namespace scoped, all
the profiles were originally named `bpfman-secure`. Now that they are
cluster scoped, they need unique names. So this PR just made the SeLinux
Profile name `bpfman-secure-<ExampleNamespace>` to keep them unique.

In v0.9.0, there was a bug in the SPO code. When the namespace was
dropped, the logic processing the SeLinux Profile did not drop the
namespace, but the namespace was empty. As a result, the code in v0.9.0
expects <SeLinuxProfileName>_.process (extra "_" at the end). This was
fixed with:
https://github.com/kubernetes-sigs/security-profiles-operator/issues/2745
https://github.com/kubernetes-sigs/security-profiles-operator/pull/2746
There is another fix bpfman needs:
https://github.com/kubernetes-sigs/security-profiles-operator/issues/2699
https://github.com/kubernetes-sigs/security-profiles-operator/pull/2746

SPO plans to cut a new release, resumably v0.9.1 with these changes.
When that happens, bpfman needs this PR. If v0.9.1 does not happen, this
PR needs to be updates to include an "_" in the form of
<SeLinuxProfileName>_.process.

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>