bpfman: invalid link fixes
If processing a `bpfman attach` failed in certain scenarios, the link
would be added to the program but not to the DB. Once that happens, the
program cannot be unloaded.

Example, load an XDP program and then try to attach to an invalid
interface.

Step one is in the attach_program() function, once `prog.add_link()` is
called, don't exit the function without calling `prog.remove_link()`.

Step two, when in the Program object and walking through the links to
build up a list of link objects, if the link is not in the DB, skip over
it because there is nothing we can about it at that point. We don't have
enough data to detach.

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>