Fix get_dispatcher() function
The issue was that the "contains" check in get_dispatcher could match
multiple dispatchers. For example, a Dispatcher ID of
xdp_dispatcher_4026531840_2 is contained in
xdp_dispatcher_4026531840_20 and xdp_dispatcher_4026531840_25.  As a
result, the code could get the wrong dispatcher causing dispatchers
to be unexpectedly deleted or hijacked, and that caused problems.
Whether it happened depended on the order in which the interfaces
were discovered and processed.

We didn't see this issue before because we never tested many
interfaces in any given network namespace.

The above bug also exposed a couple of corner cases in
detach_multi_attach_program() which were tightened up.  These cases
should never have happened, but they did because of the above bug.

Fixes: #1537

Signed-off-by: Andre Fredette <afredette@redhat.com>