Fix go-xdp-counter-sharing-map example deployment error
When running `make -C examples deploy`, the go-xdp-counter-sharing-map
example was failing with an error due to incorrect field names in the
ClusterBpfApplication resource definition.

The issue was that the program type was specified as "xdp" (lowercase)
instead of "XDP" (uppercase), and the XDP configuration was
incorrectly nested under "xdpInfo" instead of "xdp".

This change fixes the field names to match the expected
ClusterBpfApplication schema, allowing the example to deploy
successfully alongside the other examples.

Signed-off-by: Andrew McDermott <amcdermo@redhat.com>