Add nudge annotation to bpfman-daemon-ystream pipeline
Add the build-nudge-files annotation to bpfman-daemon-ystream-push.yaml
to enable Konflux nudging when this component builds. The annotation
points to hack/konflux/images/bpfman.txt which is the new location for
image reference updates in the bpfman-operator repository.

This restores the nudging functionality that was broken when the nudge
files were reorganised from hack/update_configmap.sh to
hack/konflux/images/ in the bpfman-operator repository.

When bpfman-daemon-ystream builds, Konflux will now create PRs to update
the image reference in bpfman-operator, triggering rebuilds of dependent
components as configured in the Konflux application.